Huge Update
I had to make a commit to keep track of changes and bring new things—the service.sh is not working correctly for now. I did some tests using shell, and it all worked adequately, but Magisk does not seem to do its job perfectly. I plan to add system-less themed icons and a modified Pixel Launcher on the next update.